chore(refactor): move  common decryption worker utils to commons
