+++ apps/frontend/src/features/admin-form/responses/ResponsesPage/storage/StorageResponsesService.ts
@@ -1,4 +1,3 @@
-import { releaseProxy, wrap } from 'comlink'
 import { mapValues } from 'lodash'
 
 import { API_BASE_URL } from '~services/ApiService'
@@ -6,9 +5,6 @@ import { API_BASE_URL } from '~services/ApiService'
 import { ADMIN_FORM_ENDPOINT } from '~features/admin-form/common/AdminViewFormService'
 
 import { ndjsonStream } from './utils/ndjsonStream'
-import type { DecryptionWorkerApi } from './worker/decryption.worker'
-import DecryptionWorker from './worker/decryption.worker?worker'
-import { CleanableDecryptionWorkerApi } from './types'
 
 export type EncryptedResponsesStreamParams = {
   startDate?: string
@@ -41,23 +37,4 @@ export const getEncryptedResponsesStream = async (
   })
     .then((res) => res.body)
     .then(ndjsonStream)
-}
-
-/**
- * Creates a worker, a cleanup function and returns it
- */
-export const makeWorkerApiAndCleanup = (): CleanableDecryptionWorkerApi => {
-  // Create a worker and wrap it with comlink for ease of interaction.
-  const worker = new DecryptionWorker()
-  const workerApi = wrap<DecryptionWorkerApi>(worker)
-
-  // A cleanup function that releases the comlink proxy and terminates the worker
-  const cleanup = () => {
-    workerApi[releaseProxy]()
-    worker.terminate()
-  }
-
-  const workerApiAndCleanup = { workerApi, cleanup }
-
-  return workerApiAndCleanup
 }
